p0.08-static_typing: Updated SternghCalculation to use stat collection classes.
Refactored StrengthCalculation to use OffensiveStats and DefensiveStats classes to determine the strength of an action. Adjusted type declaration for argument of get_elemental_potency in Potency. Added logic to override values for light and dark elements for OffensiveStats and DefensiveStats. Added function set_magic in OffensiveStats that updates the value of any element type, overriding light or dark as needed.