p0.08-summon: Updated Action initialization. Enabled Summon button toggling.
Adjusted Action initialization to immediately connect the potency stats for all effects to be that of the action as that should not change. Added setter logic to the source_stats parameter of Action so that the effects' source stats are automatically updated. Enabled logic that lets the Summon button toggle between other options.