p0.08-summon: Removed signal chaining and added logic to check for valid spawn action.
Refactored how a selection from a list of sub options is passed to the relevant nodes. Instead of chaining the selection signal through the relevant nodes, the relevant nodes look for the signal from the SignalBus.
Added logic that updates the wisp cost for a spawn action to match that of the summon the spawn action is for. This is to allow for the selector to be able to use existing logic to verify that a selected action can be executed.
Created initial logic to prevent the emission point of a spawn action from being placed on a tile that is already occupied. Adjusted the layout of code to keep it within readable boundaries.