p0.08-summon: Fixed bug with placing spawn action to closest tile to target.
Added logic to SelectorNodeSelectAction to track the summon of the currently selected spawn action if appropriate. Added logic to prevent from adding removed character indexes in source ranges when getting the emission selection in SelectorNodeSelectAction. Added function to SummonButton to get the name of the summon. Added function to SubOptions to get a specific SubOptionButton node.