p0.08-summon: Adjusted input mapping and fixed movement bug.
Updated the input mappings for built-in ui directional inputs (ui_left, ui_right, etc.) to not include joystick input. Adjusted the size and position of the character map hitbox to fix bug where the selector would be placed at the last tile crossed when the character ends movement above or below the previous tile.