p0.08-summon: Fixed bug with confirming spawn action to use.
FIxed a bug where given multiple summons with the same spawn action, said action would always be considered a spawn action of only the leftmost summon in the suboptions menu. Fixed this by adding a check in SelectorNodeSelectAction verifying that the summon of a spawn action matches that of the last action. Added logic to EncounterUIAction that updates the recorded selected summon when a spawn action is selected.