p0.08-summon: Added logic to place selected summon on the map.
Added signal in Selector that indicates when a spawn action has been selected, passing along the summon name and position to place it. This signal is connected to the Summon node and triggers the load_summon function. Removed checks for cardinal ranges in code that works with effects, instead relying on the emit_from_caster flag.