p0.08-summon: Began working on adding summon display logic to InitiativeTracker.
Created new signals for Summon, activated and deactivated, that indicate when the summon has been activated and deactivated respectively. Updated InitiativeSlot to have logic to change the color to indicate a summon's turn. Updated EncounterUI to disable the summon option button when a summon is active. Implemented initial version of updating the InitiativeTracker to display the summon portrait in place of the summoner's. There is currently a bug where the summoner portrait swaps between player characters as initiative progresses.