p0.08-summon: Updated setter for summoner in Summon.
Added check in setter for summoner parameter in Summon so that the summoner is not updated when the Summon is active. This prevents the stats an active summon uses from changing every time a player starts their turn. This also addresses the bug in the InitiativeTracker where the summon portrait kept changing at the start of another player's turn.