p0.08-summon: Reorganized files and added logic to handle summons to UI nodes.
Renamed folder PlayerStats to UserCharacterStats to reflect that the nodes in this file will be used to display more than just player character details. Added logic to display the battle portrait of a character instead of the default icon. Added a back panel in ActiveCharacterStats and PartyCharacterStats to handle transparent parts of the battle portrait. Created specific functions in UserCharacterStatsUI to handle summon and player data.