p0.08-summon: Moved adding of Action node as child to immediately after creating it.
Not all of the parameters of the Action node were set while it was left outside of the scene tree. Adding it into the scene tree before setting up the rest of the parameters resolves the initialization issues as the parameters that were set depended on the Action node being fully ready.